requirement.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pymongo==4.6.1
bcrypt==4.1.2
python-dotenv==1.0.0
